refactor: Update updateRoom action to include max capacity
Update the updateRoom action in the room.ts file to include the max capacity parameter. This allows for checking if the number of users exceeds the maximum room capacity. The max capacity is now passed as a parameter in the updateRoom function call in the EditRoomCard component.

Related recent commits:
- refactor: Update room data table columns and export filename
- refactor: Remove unused setSearchQuery call in EditRoomCard component