refactor: Fix user logout bug in Navbar component
The Navbar component in the components/navbar/Navbar.tsx file is updated to fix a bug related to user logout. Previously, the code was calling a non-existent function `logOut()` instead of the correct function `logout()`. This bug prevented the user from being logged out properly. The fix ensures that the correct function is called when the user is not logged in, improving the functionality of the Navbar component.

Related recent commits:
- refactor: Update middleware to handle banned users
- refactor: Update email links to use production URL
- refactor: Update LOCKOUT_DURATION and RESET_DURATION values in login.ts and sanitize user data in ProfileCard component
- refactor: Sanitize user data in ProfileCard component
- refactor: Update roomStats handling in AdminDashboard component
- refactor: Update room change approval logic and handle contract and invoice deletion
- refactor: Update login logic to handle account lockout and failed login attempts
- refactor: Update ManageRequests page to use room change requests instead of user list
- refactor: Disable delete button for pending requests in ActionsCell component
- refactor: Add banUser and unbanUser functions to admin/user module
- refactor: Update EditRoomCard dialog description for consistency
- fix student dashboard
- change build script
- Merge pull request #23 from phannhat17:student-action